All peripherals 100% supported by the cmd_dispatcher (RDA5807, ESP8266, 25AA1024)

+++ Core/cmd_dispatcher/cmd_rda5807m.c
@@ -1,6 +1,6 @@
 #include "cmd_rda5807m.h"
 
-uint16_t CmdRDA5807mDoInit(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mDoInit(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
@@ -12,7 +12,7 @@ uint16_t CmdRDA5807mDoInit(const uint8_t* const pStrCmd, const uint8_t lng) {
 	return CMD_RET_OK;
 }
 
-uint16_t CmdRDA5807mDoReset(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mDoReset(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
@@ -25,7 +25,7 @@ uint16_t CmdRDA5807mDoReset(const uint8_t* const pStrCmd, const uint8_t lng) {
 
 }
 
-uint16_t CmdRDA5807mSetMute(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mSetMute(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG*2 + CMD_ARG1_LNG+
@@ -43,7 +43,7 @@ uint16_t CmdRDA5807mSetMute(const uint8_t* const pStrCmd, const uint8_t lng) {
 	return result;
 }
 
-uint16_t CmdRDA5807mSetFreq(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mSetFreq(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG*2 + CMD_ARG5_LNG+
@@ -66,7 +66,7 @@ uint16_t CmdRDA5807mSetFreq(const uint8_t* const pStrCmd, const uint8_t lng) {
     return result;
 }
 
-uint16_t CmdRDA5807mSetVolm(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mSetVolm(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG*2 + CMD_ARG2_LNG+
@@ -87,7 +87,7 @@ uint16_t CmdRDA5807mSetVolm(const uint8_t* const pStrCmd, const uint8_t lng) {
     return result;
 }
 
-uint16_t CmdRDA5807mGetRSSI(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mGetRSSI(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
@@ -97,7 +97,7 @@ uint16_t CmdRDA5807mGetRSSI(const uint8_t* const pStrCmd, const uint8_t lng) {
 	return RDA5807mGetRSSI();
 }
 
-uint16_t CmdRDA5807mGetRDSR(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mGetRDSR(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
@@ -107,7 +107,7 @@ uint16_t CmdRDA5807mGetRDSR(const uint8_t* const pStrCmd, const uint8_t lng) {
 	return RDA5807mGetRDSR();
 }
 
-uint16_t CmdRDA5807mGetRDSS(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mGetRDSS(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
@@ -117,7 +117,7 @@ uint16_t CmdRDA5807mGetRDSS(const uint8_t* const pStrCmd, const uint8_t lng) {
 	return RDA5807mGetRDSS();
 }
 
-uint16_t CmdRDA5807mGetSTRO(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mGetSTRO(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
@@ -127,7 +127,7 @@ uint16_t CmdRDA5807mGetSTRO(const uint8_t* const pStrCmd, const uint8_t lng) {
 	return RDA5807mIsChannelStereo();
 }
 
-uint16_t CmdRDA5807mGetCHST(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mGetCHST(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
@@ -137,7 +137,7 @@ uint16_t CmdRDA5807mGetCHST(const uint8_t* const pStrCmd, const uint8_t lng) {
 	return RDA5807mIsChannelStation();
 }
 
-uint16_t CmdRDA5807mGetBLKA(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mGetBLKA(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
@@ -148,7 +148,7 @@ uint16_t CmdRDA5807mGetBLKA(const uint8_t* const pStrCmd, const uint8_t lng) {
 
 }
 
-uint16_t CmdRDA5807mGetBLKB(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mGetBLKB(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
@@ -159,7 +159,7 @@ uint16_t CmdRDA5807mGetBLKB(const uint8_t* const pStrCmd, const uint8_t lng) {
 
 }
 
-uint16_t CmdRDA5807mGetBLKC(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mGetBLKC(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
@@ -170,7 +170,7 @@ uint16_t CmdRDA5807mGetBLKC(const uint8_t* const pStrCmd, const uint8_t lng) {
 
 }
 
-uint16_t CmdRDA5807mGetBLKD(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mGetBLKD(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
@@ -181,7 +181,7 @@ uint16_t CmdRDA5807mGetBLKD(const uint8_t* const pStrCmd, const uint8_t lng) {
 
 }
 
-uint16_t CmdRDA5807mGetERRA(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mGetERRA(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
@@ -192,7 +192,7 @@ uint16_t CmdRDA5807mGetERRA(const uint8_t* const pStrCmd, const uint8_t lng) {
 
 }
 
-uint16_t CmdRDA5807mGetERRB(const uint8_t* const pStrCmd, const uint8_t lng) {
+uint16_t CmdRDA5807mGetERRB(const uint8_t* const pStrCmd, const uint16_t lng) {
 
     if ((CMD_METHOD_LNG + CMD_NAME_LNG +
          CMD_DELIMITER_LNG + CMD_EOL_LNG) != lng) {
